pyproject.toml
[tool.poetry]
name = "datadog-mcp"
version = "0.1.0"
description = ""
authors = ["andriirimovych <ar@ar.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
mcp = "^1.0.0"
pydantic = "^2.0"
datadog-api-client = "^2.30"
httpx = "^0.27"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
